+++ src/gpio.rs
@@ -280,6 +280,26 @@ impl<const P: char, const N: u8, MODE> PinExt for Pin<P, N, MODE> {
     }
 }
 
+pub trait PinSpeed {
+    /// Set pin speed
+    fn set_speed(&mut self, speed: Speed);
+}
+
+pub trait PinPull {
+    /// Set the internal pull-up and pull-down resistor
+    fn set_internal_resistor(&mut self, resistor: Pull);
+}
+
+impl<const P: char, const N: u8, MODE> PinSpeed for Pin<P, N, MODE>
+where
+    MODE: marker::OutputSpeed,
+{
+    #[inline(always)]
+    fn set_speed(&mut self, speed: Speed) {
+        self.set_speed(speed)
+    }
+}
+
 impl<const P: char, const N: u8, MODE> Pin<P, N, MODE>
 where
     MODE: marker::OutputSpeed,
@@ -299,6 +319,16 @@ where
     pub fn speed(mut self, speed: Speed) -> Self {
         self.set_speed(speed);
         self
+    }
+}
+
+impl<const P: char, const N: u8, MODE> PinPull for Pin<P, N, MODE>
+where
+    MODE: marker::Active,
+{
+    #[inline(always)]
+    fn set_internal_resistor(&mut self, resistor: Pull) {
+        self.set_internal_resistor(resistor)
     }
 }
 
@@ -464,6 +494,24 @@ impl<const P: char, const N: u8, MODE> Pin<P, N, Output<MODE>> {
         } else {
             self.set_low()
         }
+    }
+}
+
+pub trait ReadPin {
+    #[inline(always)]
+    fn is_high(&self) -> bool {
+        !self.is_low()
+    }
+    fn is_low(&self) -> bool;
+}
+
+impl<const P: char, const N: u8, MODE> ReadPin for Pin<P, N, MODE>
+where
+    MODE: marker::Readable,
+{
+    #[inline(always)]
+    fn is_low(&self) -> bool {
+        self.is_low()
     }
 }